+++ core/component.go
@@ -68,18 +68,6 @@ func RootComponentTemplate(name string) *Component {
 	}
 }
 
-// Commit ...
-func (component *Component) Commit(parent *Component) error {
-	if len(component.Jobs) == 0 {
-		return nil
-	}
-	eg := new(errgroup.Group)
-	for _, job := range component.Jobs {
-		eg.Go(job.Commit)
-	}
-	return eg.Wait()
-}
-
 // Create ...
 func (component *Component) Create() error {
